✨ Introducing typescript on NextJS
