+++ .gitignore
@@ -41,3 +41,4 @@ yarn-error.log*
 
 **/node_modules
 **/dist
+**/*.js
